tc-cloud/requirements.txt
fastapi[standard]==0.121.0
fastapi-mqtt==2.2.0
Jinja2==3.1.6
python-dotenv==1.2.1
pandas==2.3.3